requirements.txt
PySide2
requests
python-dotenv